RAG4Dify/requirements.txt
flask
python-dotenv
langchain-google-genai
pinecone
gunicorn
requests
